Change AtomicReference#initialize to accept false
Currently, Rubinius::AtomicReference#initialize will
silently discard the desired initial value for the
reference if it is 'false'. If one opts not to provide
an initial value for the reference, it defaults to nil.
In this case, the #initialize method will not set the
value of the AtomicReference to nil. However, rather
than checking if the initial value argument is nil,
it checks if it is falsey. This means that both nil
and false values are excluded. This patch modifies the
method to explicitly check if the value is nil.