Unify write barrier access
This ensures we only have one implementation for the write barrier.
Before Object had it's own implementation and the write barrier itself
as well.

Also removes the aux barrier mechanism for now, since we don't need it.
If we want / need it, we can introduce it again in the future.