Introducing concurrent mature mark phase
This creates the option of having a concurrent mark phase for the mature
GC, which is enabled by default.

The strategy is a straightforward implementation, where we first do a
stop the world and an initial scan of the roots and threads. After this
we start a concurrent mark phase.

If the mark stack has been processed concurrently, we stop the world
again and rescan the roots and threads. We also mark all objects that
have been writting into the mark set through the write barrier. After
this we mark all pending objects as a final phase.

Because we don't want to prevent a young GC while we're doing a
concurrent mark phase, we have to do some additional work for young GC
cycles if a mature mark phase is in progress. We basically scan all
mature objects in the mark stack, because they act like a remember set
and can point to young objects.

We also have to remove any unreachable young objects from the mark
stack. The weakref list needs the same treatment here during a young GC.

Another thing we do is rescan all the marked RData objects. This is
needed because there is no write barrier for those in a C extension, so
we have to make sure to rerun rb_mark() to ensure we scan them properly.
Otherwise there might be active references inside an object wrapped with
Data_Wrap_Struct we accidentally clean up.

When an object is promoted during the young GC, we explicitly clear the
current mark. This is done so we will remark it through the write
barrier, and the underlying lines in Immix will be marked for this
object. Otherwise we will think the object is marked, but the underlying
immix lines aren't and will be reused, resulting in crashes.

The marker thread basically functions like any auxiliary thread, like
the signal and finalizer threads. After a fork or before an exec, we
explicitly say the mature mark isn't in progress anymore. Mature marking
can be properly restarted then afterwards because it would never finish
and would never start a new mature cycle anymore otherwise.