We always create a code resource as marked initially
This ensures that if we create a new code resource after the object
storing it is already marked, we don't clean it up accidentally.

This can result in the code resource to be kept alive one more GC cycle
in cases where it's immediately discarded, but this is an unlikely
scenario with code resource objects.