Use GCData pointer instead of reference
This is needed because we want to manage this outside the stack later
for concurrent GC. In that case we need to store the GCData on the heap
so the concurrent thread can use it.

Therefore we replace this in advance, so the change for concurrent GC
can be smaller.