Improve young GC auto tuning
This allows for automatically growing the young generation space if
there is a lot of pressure on the young generation. This pressure is
measured to tracking how much young GC's happen while we do a concurrent
GC.

If this more than once, we check the current size of the young
generation and request a growth if the mature space is larger than the
young space by a given factor. The current default factor is 8, which is
a fairly conservative factor based on metrics obtained through
benchmarks.