Start refactoring for allowing multiple compiled regexps
This is the first step to allowing multiple compiled regexps to be
cached. For example in Rails it happens very often that regexps are
recompiled all the time. This is because they are compared with
US-ASCII, UTF-8 and ASCII-8BIT strings.

This cache will have dedicated slots for these three types and a fourth
slot for other encodings which are much rarer.