Fix thread safety issue with modifying constant scope for code
Modifying compiled code objects at runtime results in race conditions if
the code is executed under different contexts in multiple threads.

This changes it to always use the constant scope from the call frame and
set this up properly for all the cases. Also the JIT needs to setup
these frames properly.

Also ensures that the JIT also checks the constant scope when validating
an inlined constant.

Fixes #2221