Remove usage of MachineCode::parent
This was trying to keep some form of relationships between machine code
objects that was used only for JIT deopts. The problem is that is
absolutely not thread safe and can cause issues with deoptimizations.

This new approach uses the meta information already available in the JIT
to emil the call frame for the actual creator of the scope, so we can
use that. That is thread safe and doesn't have the issues of using
MachineCode::parent.

Fixes #2383