Better CPU detection
This copies the current CPU detection from LLVM trunk. There is an issue
with certain i7 CPU's where AVX is not available. One example of this is
the Linode vm's. There LLVM doesn't detect AVX is not available and
enables it.

This leads to SIGILL errors because it generates instructions the CPU
can't handle. This new detection properly handles this case since it has
additional logic for AVX and solves the issue.