Use US-ASCII as the default encoding for Symbol for now
This is not always correct, but much more compatible. For example
dumping a hash with symbol keys to YAML would break since the string was
marked as binary encoding. When retrieving it would load the original
symbol name.

We need to add proper symbol encoding support in the future, but for now
this is a better hack.