Fix handling of Rational() and Complex() when Mathn is loaded
In these cases when the denominator is 1 for Rational and the imaginery
component for Complex is 0, we return the integer value when Mathn is
loaded.

This behavior ensures that numerical division of whole numbers where
there would be no remainder work properly. This would otherwise for
example cause exceptions in the Redis driver where the return value of a
length computation is embedded in the command string sent to the server.
So if an application would load mathn, Redis would break.