Removed pre-installed ffi gems.
The current versions of the ffi gem install on Rubinius. These
gems were bundled to try to help transitioning from the built-in
Rubinius FFI to the gem. But as with most things that hide
potential problems, this is likely more trouble than good. Better
to force projects on ffi pre-1.2 to update.