Fix caching of responds_to with method missing and vm_check_callable
This is used in Rails 4.0 for determining template options, so this
would break tests in Rails.

This also adds a Rubinius specific spec we properly cache the respond_to
value. This means we have to run the respond_to_missing code from the vm
primitive itself so we cache the right value.