Don't skip waiting if no one waits yet
This was causing a race condition where the debugger thread wasn't
waiting yet, but was going to. This meant the executing thread didn't
stop and continued to run.

There is no problem here to wait, since that was already explicitly
requested. We don't know who will be first here, so just not checking
this is the best solution.

Fixes #2393