Slightly reformatted the style of backtraces.
This commit introduces two extra newlines in the format, one before the message
and one after each "Backtrace:" header (including the ones used for custom
backtraces).

This means that instead of this (where "..." would be the message):

    An exception occurred running /tmp/test.rb
        ... (NameError)

    Backtrace:
      Kernel(Module)#number (method_missing) at kernel/delta/kernel.rb:81

You get this instead:

    An exception occurred running /tmp/test.rb:

        ... (NameError)

    Backtrace:

      Kernel(Module)#number (method_missing) at kernel/delta/kernel.rb:81

It's an extremely pedantic change, partially driven by a bad case of OCD, but
it does make the error message stand out more (at least in my opinion).

Related issue: #2496.