Properly fix timeout for nested requests
This together with the revert of
97416082698c88ffdad4131db155ff152d24baeb makes sure that timeouts work
properly without spurious interrupts.

The spurious interrupts where caused by a new timeout request being
scheduled and triggering a reschedule. This would cause the sleep
timeouts to be updated, but also for the new request.

This means the new request already had time subtracted before a timeout
really happened. The fix therefore is to only add the new request after
we've handled sleep time updates.

Fixes #2499