Split out beginning of lookup and direct class for object
This also sets up an additional origin field in a class which indicates
the beginning of the lookup chain for this class. This is setup so we
can use this for implementing Module#prepend.

In a lot of places we also used lookup_begin() as a mechanism to get the
direct class for example for cache comparisons. For this case we
introduce a new better named method called direct_class() so that can be
used to get the actual class.