Implement Module#prepend
This sets up prepending. What it does for subclassing is making sure the
subclass points to the origin class, which could be an included module.
This is done so that method lookup is still simple.

First this was attempted by walking back into the origin during lookup,
but this was causing issues with infinite loops. This approach turns out
to be much simpler in it's implementation.

Implementation is also very loosely based on some experimentation done
in #1848

Fixes #1848