Rework Module#prepend and origin logic
This switches to a different mechanism to fix some issues with
Module#prepend exposed by previously added specs.

The problem was mostly wrong ordering and prepending into already
included modules into other classes. The first problem was solved by
actually using append logic for the additional modules that can be
included by a prepend.

Before all the modules from the prepended module were prepending,
resulting in a wrong ordering.

The redefinition problem is solved by making origin just a marked for
redefinition if it's different. This results in lookup logic to skip
that class / module that has a different origin and encounter it later
because it was readded as an included module later.