Add String#delete cases for benchmarks
Also move down empty check since otherwise the JIT specializes for empty
strings causing a massive skew in the performance of the other versions.
In real situations it is very unlikely that the JIT heats up for a
String#delete! with empty strings, so this gives a better real world
picture.