Also track call count increases due to looping
This allows for starting with the right number of hits when compiling a
method. Consider the follow sample code:

def foo
  i = 0
  bar(i)
  while i < 1000000
    quz(i)
  end
end

In this case the method foo heats up because of the loop it executes.
But when we JIT foo, we didn't inline the call to bar() since that was
called only once while it throught foo was called 1000000 times.

By tracking this count for loops separate, we can substract it from the
call count to initialize the hits counter for the toplevel method with
the right value. In that case it would be 1 as well as the count for
bar() and it would inline bar() properly.

This was extracted from example from how String#delete! behaves. This
uses a similar mechanism with a loop.