Retry algorithm for when sysconf() fails
This uses a default buffer of size 2048 and does a retry if this is too
small. It also throws a proper exception if any other error happens. The
default of 2048 is based on defaults also used in FreeBSD itself:

http://lists.freebsd.org/pipermail/freebsd-bugs/2011-October/046141.html

Since there is no GC possible for the getpwnam_r call, we don't have to
use a pinned ByteArray. There is no gc lock guard here or a gctoken
passed down, so it can move during this system call. We also can use a
non zeroed version since it's written to immediately anyway.