Add encoding support for Symbol
This is proably the simplest implementation based on what there is now.
In the future we want to rework the SymbolTable, but for now we can at
least support this.

Symbol encodings are used for example in the i18n gem, so there's real
code out there depending on this.