Fixed standard library stubs.
We use actual files so that they will be found from the $LOAD_PATH
if other things (like rubygems) override #require.

This also gives us the ability to provide a meaningful error message
that explains how to install the standard lib gem and what to do in
the confusing situations where a gem is installed but GEM_HOME or
GEM_PATH are messed up.