Improve performance of String#*
This changes two things. First it alloces a dirty byte array so we don't
zero the byte array storage and immediately overwrite it.

Seconds is a smarter strategy for creating the pattern by using doubling
as long as we can and then copy the final data in a single go.

Before:
=== bin/rbx ===
            String#*    58992.9 (±1.9%) i/s -     297055 in   5.037281s

After:
=== bin/rbx ===
            String#*   635072.8 (±12.3%) i/s -    3131848 in   5.015714s