Force use of staging prefix for building Melbourne. Fixes #2630.
The issue is that when you configure with --prefix, running 'rake' will install. Now
the place that Rubinius was configure to run has a complete runtime directory structure.
So when you start up rbx during the build process with a subsequent 'rake install',
rbx will find the installed dirs, except we don't install the .rbc files because these
are cached during runtime in ~/.rbx if at all.

Short story is, 'rake; rake install' is redundant. Probably need to fix rake install
to note that rbx is already installed or something.