Recheck mark stack for objects after scanning fibers
When we scan stuff like finalizers etc. it can happen we end up pushing
a new object into the mark stack due to the write barrier when
a concurrent mature mark phase is also in progress.

By rescanning this as well in the loop we make sure we handle all
objects for which this might happen.