Don't use iterator to walk vector that can grow
Problem here is that walking this mark set can result in new entries
being added. This can happen because the write barrier runs for objects
that are seen and this can add elements.

The new entries added can result in a vector resize and that invalidates
all the iterators. This way we prevent this problem.