Add back sitedir
This reverts the changes for sitedir and only keeps the changes to
rubyarchhdrdir for headers. That last one needs to point to the C-API
headers since C extensions depend on that and was the reason also for
the original change.

We should also add actually making these directories so they are
actually available. We didn't do this even before this change.