Updated all socket constants to be compatible with Ruby 2.1
There are a lot of sockets constants that were missing in Rubinius.
This commit adds all of them to harmonize with the newest version of
Ruby.

After run platform.rake a new platform.conf file will be generated with
the right values for the current operational system.

Fixes #2729, rubysl/rubysl-socket#2 and enables to fix puma/puma#393