Remove GC attempt when no more fd's are available
This logic doesn't work (anymore) because of two major reasons. One is
that GC.run(true) now only triggers starting a GC with the concurrent
GC, it doesn't wait for it to finalize a full GC cycle. This would be
fixable, but that still doesn't solve the problem because of the second
reason.

The second reason is that we don't actually run finalizers that close
IO's during the GC phase. We always run them in the finalizer thread
that only starts running after the GC. This means that even when there
are possible candidates to close with a finalizer, they won't be closed
after the GC.run(true) call.

We should investigate a better solution for this problem, or advise
people on cleaning up IO handles better in their code. Either way, this
code doesn't really do anything anymore so we can remove it.