Fixed download of https urls under proxy
The ::Proxy class is obsolete (http://ruby-doc.org/stdlib-2.0.0/libdoc/net/http/rdoc/Net/HTTP.html#method-c-Proxy), thus it didn't had the use_ssl method, which was causing the error. So, I updated the download method to use the new way (pass the proxy parameters in constructor). I'm not sure if this is the best impl, but it works.