Fix memory usage for turning long strings into bignums
This was caused because relocation of managed mp_int storage means a new
ByteArray is allocated. Because there was no opportunity for doing a GC
during the loop, it would keep growing in memory usage.

By reusing the same non managed mp_int, this memory usage is greatly
lowered. It means less allocations and reallocations, leading to stable
memory usage.