Revert back to using std::list
We current keep iterators in instance variables which are invalidated
for a deque, but not for a list. This brings back some of the
performance issues in specific cases with using std::list so we should
review this code to see if it can be done differently to perform
better.