Re-added the initialization of setproctitle.
This was removed (most likely by accident) in commit
eaaaa24c1733527f73adb2807eaba85c13d7b591. This line however *is* required for
the setproctitle() function to work.

This commit also adds some extra tests to check if the process title *actually*
changed. Since these tests shell out to `ps` (as that's the only reliable
method that I know of to get the title) they are wrapped in a platform block
that currently only runs on Darwin and Linux systems.

This fixes #2807.