MRI compatibility for rb_check_convert_type.
Due to its usage the changes are rather, well, ugly. The problem here is that a
bunch of functions use rb_check_convert_type without specifying what the
expected type should be (before this commit said type was never even used).

For now a value of `-1` as the type will ignore the extra call to `raise` if
the coercion method exists but returns an unexpected type. In the long run we
probably want to refactor this code.