fix unusual interpreter references on #! line.
lintian found several unusal interpreters and paths on #! lines.
I removed the line from darkfish.rb as no other file in that library
had a #! line.
The others I set to /usr/bin/ruby as that appear to be the most used but
perhaps /usr/bin/env ruby would be better.