Fixed setting process title.
There's a very important comment in ruby_init_setproctitle. I'll just
quote it verbatim:

  /*
   * NB: This assumes that argv has already been copied out of the
   * way. This is true for sshd, but may not be true for other
   * programs. Beware.
   */

Nothing was saving argv and since this code does this:

  argv[1] = NULL;

it was causing a segv on Mavericks when the Agent init code was
iterating argv. Why this was not segv'ing on Linux (ie running
the CI specs on Travis) is unknown.